+++ packages/multi-tenant/src/lib/getDatabaseConfig.ts
@@ -1,19 +1,19 @@
-import type { SlonikConfig } from "@dzangolab/fastify-slonik";
+import type { SlonikOptions } from "@dzangolab/fastify-slonik";
 import type { ClientConfig } from "pg";
 
-const getDatabaseConfig = (slonikConfig: SlonikConfig): ClientConfig => {
+const getDatabaseConfig = (slonikOptions: SlonikOptions): ClientConfig => {
   let clientConfig: ClientConfig = {
-    database: slonikConfig.db.databaseName,
-    user: slonikConfig.db.username,
-    password: slonikConfig.db.password,
-    host: slonikConfig.db.host,
-    port: slonikConfig.db.port,
+    database: slonikOptions.db.databaseName,
+    user: slonikOptions.db.username,
+    password: slonikOptions.db.password,
+    host: slonikOptions.db.host,
+    port: slonikOptions.db.port,
   };
 
-  if (slonikConfig.clientConfiguration?.ssl) {
+  if (slonikOptions.clientConfiguration?.ssl) {
     clientConfig = {
       ...clientConfig,
-      ssl: slonikConfig.clientConfiguration?.ssl,
+      ssl: slonikOptions.clientConfiguration?.ssl,
     };
   }